pyproject.toml
[project]
name = "shinomni"
version = "0.1.0"
description = "Collection of Multipurpose Utilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fonttools>=4.55.6",
    "humanize>=4.11.0",
    "jinja2>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pycairo>=1.27.0",
    "shiny>=1.2.1",
    "shinyswatch>=0.8.0",
    "shinywidgets>=0.4.2",
    "unicodedataplus>=16.0.0",
]
